Add tasks and suggestions updates

+++ app/services/tasks.py
@@ -6,7 +6,7 @@ from app.models.domain.tasks import Task
 from app.models.domain.users import User
 from app.models.enums.tasks import TaskOwnerType, TaskStatus
 from app.models.schemas.tasks import UserTaskCreationRequest, GroupTaskCreationRequest, TaskFilterRequest, \
-    GroupTaskFilterRequest
+    TaskUpdateRequest
 from app.services.groups import get_group_as_member
 from app.services.groups_admin import get_group_as_admin
 
@@ -21,15 +21,16 @@ def get_task(db, task_id: int):
     return task
 
 
-def get_task_with_owner_type_validation(db, task_id: int, owner_type: TaskOwnerType):
-    task = get_task(db, task_id)
-    if owner_type == TaskOwnerType.PERSONAL and tasks_repository.get_user_task_db(db, task_id):
-        return task
-    if owner_type == TaskOwnerType.GROUP and tasks_repository.get_group_task_db(db, task_id):
-        return task
+def get_task_owner_type(db, task_id: int):
+    if tasks_repository.get_user_task_db(db, task_id):
+        return TaskOwnerType.PERSONAL
+    if tasks_repository.get_group_task_db(db, task_id):
+        return TaskOwnerType.GROUP
+    if tasks_repository.get_group_task_suggestion_db(db, task_id):
+        return TaskOwnerType.GROUP_SUGGESTED
     raise HTTPException(
         status_code=status.HTTP_400_BAD_REQUEST,
-        detail="Task owner does not match"
+        detail="Task owner type not found"
     )
 
 
@@ -140,7 +141,12 @@ def get_all_tasks(db, current_user: User, request: TaskFilterRequest):
 
 
 def delete_user_task(db, current_user: User, task_id: int):
-    task = get_task_with_owner_type_validation(db, task_id, TaskOwnerType.PERSONAL)
+    task = get_task(db, task_id)
+    if get_task_owner_type(db, task_id) != TaskOwnerType.PERSONAL:
+        raise HTTPException(
+            status_code=status.HTTP_400_BAD_REQUEST,
+            detail="This task is not personal"
+        )
     user_task_db = tasks_repository.get_user_task_db(db, task_id)
     if not user_task_db or user_task_db.user_id != current_user.id:
         raise HTTPException(
@@ -153,9 +159,31 @@ def delete_user_task(db, current_user: User, task_id: int):
 
 
 def delete_group_task(db, current_user: User, task_id: int):
-    task = get_task_with_owner_type_validation(db, task_id, TaskOwnerType.GROUP)
+    task = get_task(db, task_id)
+    if get_task_owner_type(db, task_id) != TaskOwnerType.GROUP:
+        raise HTTPException(
+            status_code=status.HTTP_400_BAD_REQUEST,
+            detail="This task is not a group task"
+        )
     group_task_db = tasks_repository.get_group_task_db(db, task_id)
     get_group_as_admin(db, current_user, group_task_db.group_id)
+    tasks_repository.delete_group_task_db(db, task_id)
+    db.delete(task)
+    db.commit()
+
+
+def delete_suggested_task(db, current_user: User, task_id: int):
+    task = get_task(db, task_id)
+    if get_task_owner_type(db, task_id) != TaskOwnerType.GROUP_SUGGESTED:
+        raise HTTPException(
+            status_code=status.HTTP_400_BAD_REQUEST,
+            detail="This task is not a suggested task"
+        )
+    suggested_task_db = tasks_repository.get_group_task_suggestion_db(db, task_id)
+    if suggested_task_db.user_id != current_user.id:
+        get_group_as_admin(db, current_user, suggested_task_db.group_id)
+    else:
+        get_group_as_member(db, current_user, suggested_task_db.group_id)
     tasks_repository.delete_group_task_db(db, task_id)
     db.delete(task)
     db.commit()
@@ -223,3 +251,64 @@ def process_suggested_task(db, current_user: User, task_id: int, is_accept: bool
         db.delete(group_task_suggestion_db)
         db.delete(task)
         db.commit()
+
+
+def update_user_task(db, current_user: User, request: TaskUpdateRequest):
+    get_task(db, request.task_id)
+    if get_task_owner_type(db, request.task_id) != TaskOwnerType.PERSONAL:
+        raise HTTPException(
+            status_code=status.HTTP_400_BAD_REQUEST,
+            detail="This task is not personal"
+        )
+    user_task_db = tasks_repository.get_user_task_db(db, request.task_id)
+    if not user_task_db or user_task_db.user_id != current_user.id:
+        raise HTTPException(
+            status_code=status.HTTP_400_BAD_REQUEST,
+            detail="This task is not assigned to you"
+        )
+    if request.status == TaskStatus.SUGGESTED:
+        raise HTTPException(
+            status_code=status.HTTP_400_BAD_REQUEST,
+            detail=f"Can't change task status to {TaskStatus.SUGGESTED}"
+        )
+    tasks_repository.apply_task_update_db(db, request)
+
+
+def update_group_task(db, current_user: User, request: TaskUpdateRequest):
+    get_task(db, request.task_id)
+    if get_task_owner_type(db, request.task_id) != TaskOwnerType.GROUP:
+        raise HTTPException(
+            status_code=status.HTTP_400_BAD_REQUEST,
+            detail="This task is not a group task"
+        )
+    group_task_db = tasks_repository.get_group_task_db(db, request.task_id)
+    if group_task_db.user_id == current_user.id:
+        get_group_as_member(db, current_user, group_task_db.group_id)
+    else:
+        get_group_as_admin(db, current_user, group_task_db.group_id)
+    if request.status == TaskStatus.SUGGESTED:
+        raise HTTPException(
+            status_code=status.HTTP_400_BAD_REQUEST,
+            detail=f"Can't change task status to {TaskStatus.SUGGESTED}"
+        )
+    tasks_repository.apply_task_update_db(db, request)
+
+
+def update_suggested_task(db, current_user: User, request: TaskUpdateRequest):
+    get_task(db, request.task_id)
+    if get_task_owner_type(db, request.task_id) != TaskOwnerType.GROUP_SUGGESTED:
+        raise HTTPException(
+            status_code=status.HTTP_400_BAD_REQUEST,
+            detail="This task is not a suggested task"
+        )
+    group_task_db = tasks_repository.get_group_task_suggestion_db(db, request.task_id)
+    if group_task_db.user_id == current_user.id:
+        get_group_as_member(db, current_user, group_task_db.group_id)
+    else:
+        get_group_as_admin(db, current_user, group_task_db.group_id)
+    if request.status != TaskStatus.SUGGESTED:
+        raise HTTPException(
+            status_code=status.HTTP_400_BAD_REQUEST,
+            detail="Can change task status for suggested task"
+        )
+    tasks_repository.apply_task_update_db(db, request)
